+++ app/Http/Controllers/Api/ApiUserFollowersController.php
@@ -13,27 +13,33 @@ class UserFollowersController extends Controller
 
     public function follow(Request $request)
     {
-            $takipeden = User::all()->where('username','=',$request->takipeden)->first();
-            $takipedilen = User::all()->where('id','=',$request->takipedilen)->first();
+        $key = "a5877455-b8ac-477f-9b7f-ccb9a979f44e";
+        if ($request->get('apiKey') == $key) {
+            $takipeden = User::all()->where('username', '=', $request->takipeden)->first();
+            $takipedilen = User::all()->where('id', '=', $request->takipedilen)->first();
             $followers = new userFollowers();
 
             $followers->followed = $takipedilen->id;
             $followers->follower = $takipeden->id;
             $followers->date = date_create();
             $followers->save();
-        return response()->json(['success' =>'Takip edildi.'],200);
+            return response()->json(['success' => 'Takip edildi.'], 200);
+        }
     }
 
     public function unfollow(Request $request)
     {
-        $takipeden = User::all()->where('username','=',$request->takipeden)->first();
-        $takipedilen = User::all()->where('id','=',$request->takipedilen)->first();
+        $key = "a5877455-b8ac-477f-9b7f-ccb9a979f44e";
+        if ($request->get('apiKey') == $key) {
+            $takipeden = User::all()->where('username', '=', $request->takipeden)->first();
+            $takipedilen = User::all()->where('id', '=', $request->takipedilen)->first();
 
-        $followers = userFollowers::all()->where('followed','=',$takipedilen->id)->where('follower','=',$takipeden->id)
-            ->first();
+            $followers = userFollowers::all()->where('followed', '=', $takipedilen->id)->where('follower', '=', $takipeden->id)
+                ->first();
 
-        $followers->delete();
+            $followers->delete();
 
-        return response()->json(['success' =>'Takipten çıkıldı.'],200);
+            return response()->json(['success' => 'Takipten çıkıldı.'], 200);
+        }
     }
 }
